Part 1
Started from CleanSlate beta 6.1; further CleanSlate updates to be merged down the line.

Lots and lots of work remains TODO; the aim is to merge systematically to catch (and fix) performance issues with Tianxia content. Playing on this branch is thus not recommended, seeing as various things will be missing and/or broken until the merge is complete.

See CleanSlate rebase progress.txt for what has been done so far.